Put result-dataframe in single module and adjust result-dict name
